backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0